backend/requirements_minimal.txt
fastapi>=0.100.0
pydantic>=2.0
python-dotenv==1.0.0
mangum>=0.17.0